refactor: move the app demos into a reusable demo package
