+++ AGENTS.md
@@ -31,6 +31,13 @@ The landing page at accountant24.ai:
 - Cloudflare Workers — hosting; the Worker also proxies `/docs` to the Mintlify docs site (`docs/`)
 - PostHog — web analytics (cookieless)
 
+## packages/demo
+
+The scripted app demos the website plays:
+
+- Astro — the mock app window components
+- Tailwind CSS — styling, using the app-look theme this package ships
+
 # Architecture
 
 `packages/desktop` follows the standard electron-vite layout (`src/main`, `src/preload`, `src/renderer`), plus one addition:
@@ -44,6 +51,8 @@ The landing page at accountant24.ai:
 - `src/shared/` — IPC payload types used by both main and renderer (and the agent host). Types only, imported with `import type` from both sides; never add runtime code here.
 
 `packages/website` is the marketing site: content and copy live in `src/content/site.ts`, pure logic in `src/lib/` and `src/worker/` (unit-tested), pages and components in `.astro` files. Wording follows the README and docs; keep it to shipped features.
+
+`packages/demo` holds the scripted app demos the site plays: one scene per file under `src/scenes/`, and everything reusable under `src/shared/` (the scene format, the timing, the mock app window, the app-look theme and fonts). It is consumed from source, so a scene or a mock change needs no build step. Anything that renders the app belongs here; page copy and page behavior stay in the website.
 
 The agent itself is `packages/pi-extension`, bundled and loaded into the agent-host utilityProcess that main forks lazily (one process for all chats).
