+++ package.json
@@ -23,7 +23,7 @@
     "lint": "biome lint .",
     "format": "biome format --write .",
     "check": "biome check --write .",
-    "typecheck": "tsc --noEmit && npm run typecheck -w @accountant24/desktop && npm run typecheck -w @accountant24/website",
+    "typecheck": "tsc --noEmit && npm run typecheck -w @accountant24/desktop && npm run typecheck -w @accountant24/website && npm run typecheck -w @accountant24/demo",
     "verify": "biome check --write . && biome check . && npm run typecheck && vitest run --coverage",
     "pre-commit": "biome check --write . && biome check .",
     "release": "gh workflow run release.yml -f type=stable",
@@ -57,43 +57,43 @@
     "noAuthors": true,
     "types": {
       "feat": {
-        "title": "🚀 Features",
+        "title": "\ud83d\ude80 Features",
         "semver": "minor"
       },
       "fix": {
-        "title": "🐞 Bug Fixes",
+        "title": "\ud83d\udc1e Bug Fixes",
         "semver": "patch"
       },
       "perf": {
-        "title": "⚡ Performance",
+        "title": "\u26a1 Performance",
         "semver": "patch"
       },
       "refactor": {
-        "title": "♻️ Refactors",
+        "title": "\u267b\ufe0f Refactors",
         "semver": "patch"
       },
       "revert": {
-        "title": "⏪ Reverts",
+        "title": "\u23ea Reverts",
         "semver": "patch"
       },
       "build": {
-        "title": "📦 Build",
+        "title": "\ud83d\udce6 Build",
         "semver": "patch"
       },
       "docs": {
-        "title": "📖 Documentation"
+        "title": "\ud83d\udcd6 Documentation"
       },
       "chore": {
-        "title": "🏡 Chore"
+        "title": "\ud83c\udfe1 Chore"
       },
       "test": {
-        "title": "✅ Tests"
+        "title": "\u2705 Tests"
       },
       "style": {
-        "title": "🎨 Styles"
+        "title": "\ud83c\udfa8 Styles"
       },
       "ci": {
-        "title": "🤖 CI"
+        "title": "\ud83e\udd16 CI"
       }
     }
   },
