+++ packages/demo/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@accountant24/demo",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    "./scenes/*": "./src/scenes/*.ts",
+    "./shared/*": "./src/shared/*.ts",
+    "./shared/components/*": "./src/shared/components/*",
+    "./shared/styles/*": "./src/shared/styles/*",
+    "./shared/fonts/*": "./src/shared/fonts/*"
+  },
+  "scripts": {
+    "typecheck": "astro check"
+  },
+  "dependencies": {
+    "@fontsource/ibm-plex-mono": "5.3.0"
+  },
+  "devDependencies": {
+    "@astrojs/check": "0.9.10",
+    "astro": "7.2.4",
+    "typescript": "5.9.3"
+  }
+}
