+++ packages/website/src/layouts/Base.astro
@@ -1,4 +1,7 @@
 ---
+import satoshi400 from "@accountant24/demo/shared/fonts/satoshi-400.woff2?url";
+import satoshi700 from "@accountant24/demo/shared/fonts/satoshi-700.woff2?url";
+import satoshi900 from "@accountant24/demo/shared/fonts/satoshi-900.woff2?url";
 import "../styles/global.css";
 import { site } from "../content/site";
 
@@ -25,9 +28,9 @@ const ogImage = new URL("/og.png", site.url).href;
     <meta name="description" content={description} />
     <link rel="canonical" href={canonical} />
     {noindex && <meta name="robots" content="noindex" />}
-    <link rel="preload" href="/fonts/satoshi-400.woff2" as="font" type="font/woff2" crossorigin />
-    <link rel="preload" href="/fonts/satoshi-700.woff2" as="font" type="font/woff2" crossorigin />
-    <link rel="preload" href="/fonts/satoshi-900.woff2" as="font" type="font/woff2" crossorigin />
+    <link rel="preload" href={satoshi400} as="font" type="font/woff2" crossorigin />
+    <link rel="preload" href={satoshi700} as="font" type="font/woff2" crossorigin />
+    <link rel="preload" href={satoshi900} as="font" type="font/woff2" crossorigin />
     <link rel="icon" type="image/png" href="/favicon.png" />
     <link rel="apple-touch-icon" href="/apple-touch-icon.png" />
     <link rel="sitemap" href="/sitemap-index.xml" />
